Change bill calculation to use integers instead of floats

+++ household_utility_analyzer/household_utility_analyzer.py
@@ -8,31 +8,32 @@ print("Billing Criteria:\nFirst 100 units = ₹5 per unit\nNext 100 units (101 t
 units = input("\nEnter the no. of units consumed by your household this month: ")
 
 try:
-    units = float(units)
+    units = int(units)
 except ValueError:
     units = False
     while units == False:
         print("\nAlphabetical or special characters were entered instead of number. Plese enter the asked info again below.")
         units = input("Enter the no. of units consumed by your household this month: ")
         try:
-            units = float(units)
+            units = int(units)
         except ValueError:
             units = False
 
 if units == 0:
     bill = 0
 elif units > 0 and units <= 100:
-    bill = units * 5.0
+    bill = units * 5
 elif units > 100 and units <= 200:
-    bill = (100 * 5) + ((units - 100) * 7.0)
+    bill = (100 * 5) + ((units - 100) * 7)
 else:
-    bill = (100 * 5) + (100 * 7) + ((units - 200) * 10.0)
+    bill = (100 * 5) + (100 * 7) + ((units - 200) * 10)
 
 print(f"\nNet demand: ₹{bill:.2f}")
 print(f"Your bill after applying a discount of 5%: ₹{bill - (0.05 * bill):.2f}")
 
 bills = np.array([])
 bills = np.append(bills, bill)
+
 file_path = r"C:\Python\household_utility_analyzer\bills.txt"
 with open(file_path, "w") as f:
     c = 0
